move interceptor options into their interceptor packages
