+++ pkg/interceptors/prestopsleep/options.go
@@ -1,0 +1,5 @@
+package prestopsleep
+
+type Options struct {
+	Seconds int `yaml:"seconds"`
+}
