+++ pkg/interceptors/statuschecker/options.go
@@ -1,0 +1,6 @@
+package statuschecker
+
+type Options struct {
+	TargetURLRegex string `yaml:"targetUrlRegex"`
+	JobRegex       string `yaml:"jobRegex"`
+}
