+++ pkg/settings/interceptors.go
@@ -1,5 +1,10 @@
 package settings
 
+import (
+	"github.com/rebuy-de/kubernetes-deployment/pkg/interceptors/prestopsleep"
+	"github.com/rebuy-de/kubernetes-deployment/pkg/interceptors/statuschecker"
+)
+
 type Interceptors struct {
 	PreStopSleep        PreStopSleepInterceptor    `yaml:"preStopSleep"`
 	RemoveResourceSpecs Interceptor                `yaml:"removeResourceSpecs"`
@@ -11,60 +16,12 @@ type Interceptor struct {
 	Enabled TriState `yaml:"enabled"`
 }
 
-type PreStopSleepOptions struct {
-	Seconds int `yaml:"seconds"`
-}
-
 type PreStopSleepInterceptor struct {
-	Enabled TriState            `yaml:"enabled"`
-	Options PreStopSleepOptions `yaml:"options"`
-}
-
-type GHStatusCheckerOptions struct {
-	TargetURLRegex string `yaml:"targetUrlRegex"`
-	JobRegex       string `yaml:"jobRegex"`
+	Enabled TriState             `yaml:"enabled"`
+	Options prestopsleep.Options `yaml:"options"`
 }
 
 type GHStatusCheckerInterceptor struct {
-	Enabled TriState               `yaml:"enabled"`
-	Options GHStatusCheckerOptions `yaml:"options"`
-}
-
-type TriState int
-
-const (
-	Unknown TriState = iota
-	Disabled
-	Enabled
-)
-
-func (ts TriState) MarshalYAML() (interface{}, error) {
-	switch ts {
-	case Enabled:
-		return true, nil
-	case Disabled:
-		return false, nil
-	default:
-		return nil, nil
-	}
-}
-
-func (ts *TriState) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var original bool
-
-	err := unmarshal(&original)
-	if err != nil {
-		return err
-	}
-
-	switch original {
-	case true:
-		(*ts) = Enabled
-	case false:
-		(*ts) = Disabled
-	default:
-		(*ts) = Unknown
-	}
-
-	return nil
+	Enabled TriState              `yaml:"enabled"`
+	Options statuschecker.Options `yaml:"options"`
 }
